Fixed Linux fonts, moved GameCard component to its own jsx file.
