+++ site/src/components/layout/PropBetContainer/PropBetContainer.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import PlayerBlock from "../../features/PlayerBlock.jsx";
 import { fetchCurrentGames, fetchGameOdds } from "../../../services/api.jsx";
 import useWindowSize from '../../../hooks/useWindowSize';
+import GameCard from "./GameCard.jsx";
 
 function PropBetContainer() {
     const [games, setGames] = useState([]);
@@ -231,36 +232,5 @@ function PropBetContainer() {
         </div>
     );
 }
-
-const GameCard = ({ game, isSelected, onClick }) => (
-    <div
-        style={{
-            border: '1px solid #ddd',
-            borderRadius: '8px',
-            padding: '15px',
-            cursor: 'pointer',
-            backgroundColor: isSelected ? 'lightgreen' : 'darkgray',
-            transition: 'all 0.2s',
-            ':hover': {
-                boxShadow: '0 2px 8px rgba(0,0,0,0.1)'
-            }
-        }}
-        onClick={onClick}
-    >
-        <p style={{ fontWeight: 'bold', margin: '0 0 5px 0' }}>
-            {game.home_team} vs {game.away_team}
-        </p>
-        <p style={{ color: '#666', fontSize: '0.9em', margin: 0 }}>
-            {new Date(game.commence_time).toLocaleString(undefined, {
-                timeZoneName: 'short',
-                weekday: 'short',
-                month: 'short',
-                day: 'numeric',
-                hour: 'numeric',
-                minute: '2-digit'
-            })}
-        </p>
-    </div>
-);
 
 export default PropBetContainer;
